Taxonomy

+++ app/Observers/FieldObserver.php
@@ -12,6 +12,7 @@
 namespace App\Observers;
 
 use App\Models\Field;
+use Illuminate\Support\Str;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 
@@ -28,9 +29,10 @@ class FieldObserver
         $fieldset   = $field->section->fieldset;
         $containers = $this->getFieldsettables($fieldset);
         
-        $fieldtype = fieldtypes()->get($field->type);
-        $column    = $fieldtype->getColumn('type');
-        $settings   = $fieldtype->getColumn('settings') ?? [];
+        $fieldtype    = fieldtypes()->get($field->type);
+        $relationship = $fieldtype->getRelationship();
+        $column       = $fieldtype->getColumn('type');
+        $settings     = $fieldtype->getColumn('settings') ?? [];
 
         array_unshift($settings, $field->handle);
 
@@ -40,6 +42,22 @@ class FieldObserver
                     call_user_func_array([$table, $column], $settings)->nullable();
                 });
             });
+        }
+
+        if (! is_null($relationship)) {
+            switch($relationship) {
+                case 'morphToMany':
+                    $tableName = Str::plural(Str::addAbleSuffix($field->handle));
+
+                    if (! Schema::hasTable($tableName)) {
+                        Schema::create($tableName, function($table) use ($field) {
+                            $table->integer($field->handle.'_id')->unsigned();
+                            $table->morphs(Str::addAbleSuffix($field->handle));
+                        });
+                    }
+
+                    break;
+            }
         }
     }
 
@@ -64,20 +82,50 @@ class FieldObserver
             $table = $container->getTable();
 
             if ($old['handle'] !== $new['handle']) {
-                $fieldtype = fieldtypes()->get($new['type']);
-                $column    = $fieldtype->getColumn('type');
+                $fieldtype    = fieldtypes()->get($new['type']);
+                $column       = $fieldtype->getColumn('type');
+                $relationship = $fieldtype->getRelationship();
 
                 if (! is_null($column)) {
                     Schema::table($table, function ($table) use ($old, $new) {
                         $table->renameColumn("`{$old['handle']}`", "`{$new['handle']}`");
                     });
                 }
+
+                if (! is_null($relationship)) {
+                    switch($relationship) {
+                        case 'morphToMany':
+                            $oldTableName = Str::plural(Str::addAbleSuffix($old['handle']));
+                            $newTableName = Str::plural(Str::addAbleSuffix($new['handle']));
+
+                            // Coooooool
+                            // https://github.com/laravel/framework/issues/2979
+                            
+                            if (! Schema::hasTable($newTableName)) {
+                                Schema::rename($oldTableName, $newTableName);
+
+                                Schema::table($newTableName, function($table) use ($old, $new) {
+                                    $table->renameColumn($old['handle'].'_id', $new['handle'].'_id');
+                                });
+
+                                Schema::table($newTableName, function($table) use ($old, $new) {
+                                    $table->renameColumn(Str::addAbleSuffix($old['handle']).'_id', Str::addAbleSuffix($new['handle']).'_id');
+                                });
+
+                                Schema::table($newTableName, function($table) use ($old, $new) {
+                                    $table->renameColumn(Str::addAbleSuffix($old['handle']).'_type', Str::addAbleSuffix($new['handle']).'_type');
+                                });
+                            }
+
+                            break;
+                    }
+                }
             }
 
             if ($oldFieldtype !== $newFieldtype) {
                 $fieldtype = fieldtypes()->get($new['type']);
                 $column    = $fieldtype->getColumn('type');
-                $settings   = $fieldtype->getColumn('settings') ?? [];
+                $settings  = $fieldtype->getColumn('settings') ?? [];
 
                 array_unshift($settings, $new['handle']);
 
@@ -100,16 +148,36 @@ class FieldObserver
     {
         $fieldset   = $field->section->fieldset;
         $containers = $this->getFieldsettables($fieldset);
+        
+        $fieldtype    = fieldtypes()->get($field->type);
+        $relationship = $fieldtype->getRelationship();
+        $column       = $fieldtype->getColumn('type');
+        $settings     = $fieldtype->getColumn('settings') ?? [];
 
-        $containers->each(function($container) use ($field) {
-            $table = $container->getTable();
+        array_unshift($settings, $field->handle);
 
-            if (Schema::hasColumn($table, $field->handle)) {
-                Schema::table($table, function ($table) use ($field) {
-                    $table->dropColumn($field->handle);
-                });
+        if (! is_null($column)) {
+            $containers->each(function($container) use ($field) {
+                $table = $container->getTable();
+
+                if (Schema::hasColumn($table, $field->handle)) {
+                    Schema::table($table, function ($table) use ($field) {
+                        $table->dropColumn($field->handle);
+                    });
+                }
+            });
+        }
+
+        if (! is_null($relationship)) {
+            switch($relationship) {
+                case 'morphToMany':
+                    $tableName = Str::plural(Str::addAbleSuffix($field->handle));
+
+                    Schema::dropIfExists($tableName);
+
+                    break;
             }
-        });
+        }
     }
 
     /**
